requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
psycopg2-binary==2.9.9
pymongo==4.6.0
neo4j==5.14.0
SQLAlchemy==2.0.23
alembic==1.13.0
python-multipart==0.0.6
redis==5.0.1